requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
alembic==1.13.3
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.12
pydantic[email]==2.9.2
